Massive Code Cleanup
Chris and I cleaned up the app and many of the recent features we've
been discussing are now included. Please hammer on it and try to break
it and let me know if you find any bugs. I deleted A LOT of code so it
is possible that something that was once working is no longer. Try
clicking things in different orders, etc. and let me know if you find
anything unexpected. Have not had a chance to test on mobile so do that
too if you can. New ideas are always welcome too!

http://maps.umn.edu/solar/

Features now live:
-EUSA and other vector layers to file geodatabase in 3857
-Bare earth disclaimer
-Out-of-state error checking
-x close on splash screen
-show/hide divs issues fixed
-UMN logo with link
-favicon